get android quest 2 build "working"
