+++ examples/demo/Cargo.toml
@@ -3,6 +3,9 @@ name = "demo"
 version = "0.1.0"
 edition = "2021"
 
+
+[lib]
+crate-type = ["cdylib"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
